requirements.txt
Flask
ultralytics
Pillow
